requirements.txt
Django==3.1.6
djangorestframework==3.12.2
pytz==2019.2
pandas==1.4.2
numpy==1.18.5